style(util): rename all util files from *_utils -> *_util
