+++ packages/util/src/_function-util.ts
