+++ packages/util/src/index.ts
@@ -1,8 +1,8 @@
-export * from "./_function-utils";
-export * from "./_js-utils";
+export * from "./_function-util";
+export * from "./_js-util";
 export * from "./_misc-types";
 export * from "./_react-types";
-export * from "./_react-utils";
+export * from "./_react-util";
 
 export * from "lodash-es";
 export * from "utility-types";
